GH-69 :: refactoring and renaming
