+++ src/TrainingGuides.Entities/ReusableContentTypes/GeneralArticle/GeneralArticle.generated.cs
@@ -65,8 +65,8 @@ namespace TrainingGuides
 
 
 		/// <summary>
-		/// ArticleSchemaTaxonomy.
+		/// ArticleSchemaCategory.
 		/// </summary>
-		public IEnumerable<TagReference> ArticleSchemaTaxonomy { get; set; }
+		public IEnumerable<TagReference> ArticleSchemaCategory { get; set; }
 	}
 }
