+++ src/TrainingGuides.Entities/ReusableContentTypes/Interview/Interview.generated.cs
@@ -35,9 +35,9 @@ namespace TrainingGuides
 
 
 		/// <summary>
-		/// IntervieweeName.
+		/// InterviewRespondantName.
 		/// </summary>
-		public string IntervieweeName { get; set; }
+		public string InterviewRespondantName { get; set; }
 
 
 		/// <summary>
@@ -71,8 +71,8 @@ namespace TrainingGuides
 
 
 		/// <summary>
-		/// ArticleSchemaTaxonomy.
+		/// ArticleSchemaCategory.
 		/// </summary>
-		public IEnumerable<TagReference> ArticleSchemaTaxonomy { get; set; }
+		public IEnumerable<TagReference> ArticleSchemaCategory { get; set; }
 	}
 }
