+++ src/TrainingGuides.Web/App_Data/CIRepository/@global/cms.class/cms.contentitemcommondata.xml
@@ -97,10 +97,10 @@
           <SelectionType>contentTypes</SelectionType>
         </settings>
       </field>
-      <field allowempty="true" column="ArticleSchemaTaxonomy" columnprecision="0" columntype="taxonomy" enabled="true" guid="7a10c083-cc91-4cf0-88af-66f580c90aa3" visible="true">
+      <field allowempty="true" column="ArticleSchemaCategory" columnprecision="0" columntype="taxonomy" enabled="true" guid="7a10c083-cc91-4cf0-88af-66f580c90aa3" visible="true">
         <properties>
           <explanationtextashtml>False</explanationtextashtml>
-          <fieldcaption>Tags</fieldcaption>
+          <fieldcaption>Categories</fieldcaption>
           <fielddescriptionashtml>False</fielddescriptionashtml>
           <kxp_schema_identifier>c3b4896f-ba7c-4b75-9cd4-47afa7489ff1</kxp_schema_identifier>
         </properties>
