+++ src/TrainingGuides.Web/App_Data/CIRepository/@global/cms.contentitemcommondata/beanplanthusbandry..jacktrott-52rzr7c8@a9141e9f85/296dd84a-59e2-4191-94f0-3fc7bf837d82_en@f1d75cde3b.xml
@@ -1,11 +1,11 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <cms.contentitemcommondata>
+  <ArticleSchemaCategory>
+    <![CDATA[[{"Identifier":"34ebdcf5-20d6-4ecc-90e6-fc150dd5b36f"}]]]>
+  </ArticleSchemaCategory>
   <ArticleSchemaSummary>
     <![CDATA[Exclusive interview with Jack Trott, a well-known figure in the world of bean plant husbandry.]]>
   </ArticleSchemaSummary>
-  <ArticleSchemaTaxonomy>
-    <![CDATA[[{"Identifier":"34ebdcf5-20d6-4ecc-90e6-fc150dd5b36f"}]]]>
-  </ArticleSchemaTaxonomy>
   <ArticleSchemaTeaser>
     <![CDATA[[{"Identifier":"e97fb6c2-06cf-4b82-bc92-ff4cf3dcd37c"}]]]>
   </ArticleSchemaTeaser>
