+++ src/TrainingGuides.Web/App_Data/CIRepository/@global/cms.contentitemcommondata/sheepcare_aninterv..lebo-peep-igkjtmw8@1140809b72/ccfcc109-5789-4bf9-99e4-904726433c95_en@c00224ac3b.xml
@@ -1,11 +1,11 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <cms.contentitemcommondata>
+  <ArticleSchemaCategory>
+    <![CDATA[[{"Identifier":"04ab8b7e-f25b-4d3e-9e78-bf319a5f9140"}]]]>
+  </ArticleSchemaCategory>
   <ArticleSchemaSummary>
     <![CDATA[Exclusive interview with Little Bo-Peep, a renowned shepherdess known for her unique approach to sheep care.]]>
   </ArticleSchemaSummary>
-  <ArticleSchemaTaxonomy>
-    <![CDATA[[{"Identifier":"04ab8b7e-f25b-4d3e-9e78-bf319a5f9140"}]]]>
-  </ArticleSchemaTaxonomy>
   <ArticleSchemaTeaser>
     <![CDATA[[{"Identifier":"4e50366d-a316-4ec0-8909-c97868aa7597"}]]]>
   </ArticleSchemaTeaser>
