+++ src/TrainingGuides.Web/App_Data/CIRepository/TrainingGuidesPages/cms.contentitemcommondata/home-ciuaqx9l@e720e8f8d0/bae62ea1-c2f8-43c6-852f-680cb2645724_en@e377933b34.xml
@@ -12,7 +12,7 @@
   </ContentItemCommonDataContentLanguageID>
   <ContentItemCommonDataGUID>bae62ea1-c2f8-43c6-852f-680cb2645724</ContentItemCommonDataGUID>
   <ContentItemCommonDataIsLatest>True</ContentItemCommonDataIsLatest>
-  <ContentItemCommonDataLastPublishedWhen>2024-08-19 20:34:25Z</ContentItemCommonDataLastPublishedWhen>
+  <ContentItemCommonDataLastPublishedWhen>2024-08-29 17:25:10Z</ContentItemCommonDataLastPublishedWhen>
   <ContentItemCommonDataPageBuilderWidgets>
     <![CDATA[{"editableAreas":[{"identifier":"areaMain","sections":[{"identifier":"b23942e8-ea2f-4754-9c81-b221abee820c","type":"TrainingGuides.SingleColumnSection","properties":{"sectionAnchor":null},"zones":[{"identifier":"d8798668-95ef-4ef4-b4fb-ed7480c4a161","widgets":[{"identifier":"3ffd6e80-6eda-4eb2-ac26-55f8d5dcba61","type":"Kentico.Widget.RichText","variants":[{"identifier":"24ca3365-1c4b-468d-8515-389ad1400aca","properties":{"content":"<p>Hello world!</p>"},"fieldIdentifiers":{"content":"cc1f8cc3-c9bf-4042-aeda-45d63e955f84"}}]}]}],"fieldIdentifiers":{"sectionAnchor":"6a4bb2cf-9e6c-4d16-b9e8-6c33ca63a9ae"}}]}]}]]>
   </ContentItemCommonDataPageBuilderWidgets>
